Fixed datatype in the profile loading

+++ implementation/gps-sleeper/src/pt/isec/gps/g22/sleeper/dal/ProfileDAOImpl.java
@@ -63,7 +63,7 @@ public class ProfileDAOImpl implements ProfileDAO {
 
         profile.setId(cursor.getInt(0));
         profile.setGender(cursor.getInt(1));
-        profile.setDateOfBirth(cursor.getInt(2));
+        profile.setDateOfBirth(cursor.getLong(2));
         profile.setFirstHourOfTheDay(cursor.getInt(3));
         return profile;
     }
